requirements.txt
discord
python-dotenv
emojis
